Fix Get select_statement position type

+++ lib/message_store/postgres/get/last/select_statement.rb
@@ -18,7 +18,7 @@ module MessageStore
               SELECT
                 id::varchar,
                 stream_name::varchar,
-                position::int,
+                position::bigint,
                 type::varchar,
                 global_position::bigint,
                 data::varchar,
